requirements.txt
python-telegram-bot==21.6
pymongo==4.6.3
APScheduler==3.10.4
python-dotenv==1.0.1
Flask==3.0.2
